feat(code_lut): embed SignalLUT in the signals added on master
Apply PR #69's signal-struct change to every signal that landed on master
after the PR branched: replace the `negated_codes` cache (retired with the
legacy gen_code! path) with a concrete `lut::SignalLUT` field and build it in
each constructor. Covers GPSL5Q, GPSL2CM, GPSL2CL, GalileoE1C, GalileoE1C_BOC11,
GalileoE5aI, GalileoE5aQ. Factor out `_gpsl5q_secondary_code` /
`_galileo_e5a_i_secondary_code` so their constructors can build the LUT before
an instance exists (mirroring `_gpsl5i_secondary_code`).

Extend the CBOC adapter to honour `boc2_sign`: bake `a1·BOC(m1,1) + a2·BOC(m2,1)`
with a negative second amplitude for the anti-phase CBOC(−) used by Galileo E1C
(the PR only handled E1B's CBOC(+)).

Port the new-signal tests to the Int8 embedded-LUT gen_code! API: drop the
`negated_codes` assertion, switch gen_code! buffers to Int8, and compare the
Int8 replica's SIGN to get_code for E1C (the CBOC bake is a scaled multi-level
approximation) — sampling at the sub-chip rate so the DDA matches get_code's
floor bit-for-bit.

+++ test/galileo/e5a.jl
@@ -109,7 +109,9 @@ end
         # Cover several primary periods so the secondary overlay is exercised.
         n_chips = 3 * get_code_length(signal)
         n_samples = samples_per_chip * n_chips
-        buf = zeros(Int16, n_samples)
+        # The embedded LUT emits Int8 ±1; at an integer samples-per-chip rate it is
+        # byte-exact to the tiered `get_code` chips below.
+        buf = zeros(Int8, n_samples)
         gen_code!(buf, signal, 1, samples_per_chip * code_freq, code_freq, 0.0, 0)
 
         @test all(x -> x == 1 || x == -1, buf)
